Add p9a_mss_freq_system procedure
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76023
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Dev-Ready: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76104

+++ src/import/chips/p9a/procedures/hwp/memory/p9a_mss_eff_config.C
@@ -64,7 +64,7 @@ fapi2::ReturnCode p9a_mss_eff_config( const fapi2::Target<fapi2::TARGET_TYPE_MEM
     {
         uint8_t l_dimm_index = 0;
         uint64_t l_freq = 0;
-        uint64_t l_omi_freq = 0;
+        uint32_t l_omi_freq = 0;
         FAPI_TRY( mss::attr::get_freq(mss::find_target<fapi2::TARGET_TYPE_MEM_PORT>(dimm), l_freq) );
         FAPI_TRY( mss::convert_ddr_freq_to_omi_freq(mss::find_target<fapi2::TARGET_TYPE_MEM_PORT>(dimm), l_freq, l_omi_freq));
